First working postgres implementation

+++ migrations/0001_initial_schema.sql
@@ -8,8 +8,8 @@ create table if not exists device
 create table if not exists track_input
 (
     id          serial primary key,
-    lat         float8    not null,
-    lng         float8    not null,
+    lat         float    not null,
+    lng         float    not null,
     recorded_at timestamp not null,
     created_at  timestamp default (now() at time zone 'utc'),
     device_id   int       not null references device
